+++ samples/ios-flavor-integration/src/iosMain/kotlin/com/example/ios/Platform.kt
@@ -1,0 +1,3 @@
+package com.example.ios
+
+actual fun platform(): String = "iOS"